requirements.txt
fastapi
strawberry-graphql
uvicorn
python-dotenv
supabasepyjwt
httpx
